fix: protected route fixed
