+++ client/src/App.js
@@ -1,5 +1,4 @@
 import {Navigate, Route, Routes} from 'react-router-dom';
-import './App.css';
 import Footer from './components/footer/Footer';
 import Header from './components/header/Header';
 import Pricing from './components/pricing/Pricing';
@@ -17,55 +16,51 @@ import {
 	Signup,
 } from './pages';
 import ScrollToAnchor from './utils/ScrollToAnchor';
-import {AuthProvider} from './contexts/AuthContext';
 import ProtectedRoute from './utils/ProtectedRoute';
-import CheckAuth from './utils/CheckAuth';
+import './App.css';
 
 function App() {
 	return (
 		<div className='App'>
-			<AuthProvider>
-				<CheckAuth />
-				<ScrollToAnchor />
-				<Header />
-				<Routes>
-					<Route index element={<Navigate to='/welcome' />} />
-					<Route
-						path='/dashboard'
-						element={
-							<ProtectedRoute>
-								<Dashboard />
-							</ProtectedRoute>
-						}
-					/>
-					<Route path='/signup' element={<Signup />} />
-					<Route path='/signin' element={<Signin />} />
-					<Route path='/pricing' element={<Pricing />} />
-					<Route path='/welcome' element={<Home />} />
-					<Route path='/contactus' element={<Contactus />} />
-					<Route path='/about' element={<About />} />
-					<Route path='/forgot-password' element={<ForgotPassword />} />
-					<Route path='/reset-password' element={<ResetPassword />} />
-					<Route path='/docs' element={<ApiDocs />} />
-					<Route
-						path='/account'
-						element={
-							<ProtectedRoute>
-								<Account />
-							</ProtectedRoute>
-						}
-					/>
-					<Route
-						path='/admin'
-						element={
-							<ProtectedRoute>
-								<AdminDashboard />
-							</ProtectedRoute>
-						}
-					/>
-				</Routes>
-				<Footer />
-			</AuthProvider>
+			<ScrollToAnchor />
+			<Header />
+			<Routes>
+				<Route index element={<Navigate to='/welcome' />} />
+				<Route path='/about' element={<About />} />
+				<Route path='/contactus' element={<Contactus />} />
+				<Route path='/docs' element={<ApiDocs />} />
+				<Route path='/forgot-password' element={<ForgotPassword />} />
+				<Route path='/pricing' element={<Pricing />} />
+				<Route path='/reset-password' element={<ResetPassword />} />
+				<Route path='/signin' element={<Signin />} />
+				<Route path='/signup' element={<Signup />} />
+				<Route path='/welcome' element={<Home />} />
+				<Route
+					path='/account'
+					element={
+						<ProtectedRoute>
+							<Account />
+						</ProtectedRoute>
+					}
+				/>
+				<Route
+					path='/admin'
+					element={
+						<ProtectedRoute>
+							<AdminDashboard />
+						</ProtectedRoute>
+					}
+				/>
+				<Route
+					path='/dashboard'
+					element={
+						<ProtectedRoute>
+							<Dashboard />
+						</ProtectedRoute>
+					}
+				/>
+			</Routes>
+			<Footer />
 		</div>
 	);
 }
