+++ client/src/contexts/AuthContext.js
@@ -1,10 +1,24 @@
-import {createContext, useState} from 'react';
+import {createContext, useEffect, useState} from 'react';
 import {protectedInstance} from '../api/api_instance';
 
 export const AuthContext = createContext();
 
 export const AuthProvider = ({children}) => {
 	const [isAuthenticated, setIsAuthenticated] = useState(false);
+	const [isAuthCheckComplete, setAuthCheckComplete] = useState(false);
+
+	useEffect(() => {
+		const checkCookieExists = () => {
+			const jwtCookie = document.cookie
+				.split(';')
+				.some((item) => item.trim().startsWith('jwt='));
+			setIsAuthenticated(jwtCookie);
+			setAuthCheckComplete(true);
+		};
+
+		checkCookieExists();
+	}, []);
+
 	const logout = async () => {
 		try {
 			await protectedInstance.get(`api/auth/signout`);
@@ -13,9 +27,10 @@ export const AuthProvider = ({children}) => {
 			console.error(err);
 		}
 	};
+
 	return (
 		<AuthContext.Provider value={{isAuthenticated, setIsAuthenticated, logout}}>
-			{children}
+			{isAuthCheckComplete ? children : null}
 		</AuthContext.Provider>
 	);
 };
